move VTypes to sigmastate.verified package;
