+++ contract-verification/src/main/scala/sigmastate/verification/contract/CrowdFundingContractVerification.scala
@@ -1,6 +1,6 @@
 package sigmastate.verification.contract
 
-import sigmastate.verification.SigmaDsl.api.sigma.{Box, Context, SigmaContract, SigmaProp}
+import sigmastate.verified._
 import stainless.annotation.ignore
 import stainless.lang._
